CEFAT4Cities/requirements.txt
pydantic>=1.9.0
translate-toolkit[XML] >=3.5.3